modified K8s-API-Servers-restart.sh script

+++ K8s-deployment/misc/K8s-ServersRestart-Script/K8s-API-Servers-restart.sh
@@ -1,43 +1,46 @@
 #!/bin/bash
 
-
 i=0
-#reads all namespaces listed in endpoint.json file
 
+read -p "you are going to restart all API servers. Continue (yes/no): " option
+#reads all namespaces listed in endpoint.json file
 cat $1 | jq -r .namespaces[] | while read ns ;
 do 
-        kubectl delete po -n $ns --all --dry-run=client
-        if [ $? == 0 ];then
+     if [ $option == "yes" ];then
+         echo
+	 #command to restart all pods in a namespace
+         kubectl delete po -n $ns --all 
+         if [ $? == 0 ];then
                 echo -e "\033[31m pods in $ns namespace are deleted\033[0m"
                 echo waiting for pods to restart....
-                sleep 30
-        fi
-        if [ $ns == "lip" ] || [ $ns == "auditing" ];then
                 sleep 60
-                ((i++))
-                continue
-        else
-        #reads all API server endpoints listed in endpoint.json file to check endpoint's health
+         fi
+         #assigns each endpoint address to endpoint var to check zero or non-zero status 
+         endpoint=$(cat $1 | jq -r .endpoints[$i])
+         if [ -z $endpoint ];then
+                  ((i++))
+                  continue
+                 
+         else
+                  status=$(curl -Lso /dev/null -m 30 $endpoint -w '%{http_code}\n')
+                  if [ $status == 200 ];then
+                    echo -e "\033[32m ------------------------------------------------------\033[0m"
+                    echo -e "\033[32m   ✅ $endpoint is reachable  \033[0m" 
+                    echo -e "\033[32m ------------------------------------------------------\033[0m"
+                    echo 
+                    ((i++))
+                    rm -rf endpoint.txt
+                    continue
+                  else 
+                    echo -e "\033[31m⚠⚠⚠\033[0m" Please recheck endpoint "\033[31m$endpoint\033[0m" address!
+                    echo -e "\033[31m⚠⚠⚠\033[0m"  Terminating script!
+                    echo
+                    rm -rf endpoint.txt
+                    break
+                 fi
 
-        cat $1 | jq -r .endpoints[$i] > endpoint.txt
-        while read endpoint; 
-        do
-                status=$(curl -Lso /dev/null -m 30 $endpoint -w '%{http_code}\n')
-                if [ $status == 200 ];then
-                 echo -e "\033[32m ------------------------------------------------------\033[0m"
-                 echo -e "\033[32m   ✅ $endpoint is reachable  \033[0m" 
-                 echo -e "\033[32m ------------------------------------------------------\033[0m"
-                 echo 
-                 ((i++))
-		 rm -rf endpoint.txt
-                 break
-                else 
-                 echo -e "\033[31m⚠⚠⚠\033[0m" Please recheck endpoint "\033[32m$endpoint\033[0m" address!
-                 echo -e "\033[31m⚠⚠⚠\033[0m"  Terminating script!
-                 echo
-		 rm -rf endpoint.txt
-                 break 2
-                fi
-        done < endpoint.txt
-        fi
-done  
+         fi
+     else
+        break
+     fi
+done
